fixing almost all windows project tests

+++ mlflow/projects/submitted_run.py
@@ -79,10 +79,13 @@ class LocalSubmittedRun(SubmittedRun):
             # Kill the the process tree rooted at the child if it's the leader of its own process
             # group, otherwise just kill the child
             try:
-                if self.command_proc.pid == os.getpgid(self.command_proc.pid):
-                    os.killpg(self.command_proc.pid, signal.SIGTERM)
-                else:
+                if (os.name == 'nt'):
                     self.command_proc.terminate()
+                else:
+                    if self.command_proc.pid == os.getpgid(self.command_proc.pid):
+                        os.killpg(self.command_proc.pid, signal.SIGTERM)
+                    else:
+                        self.command_proc.terminate()
             except OSError:
                 # The child process may have exited before we attempted to terminate it, so we
                 # ignore OSErrors raised during child process termination
